requirements.txt
pyrogram
tgcrypto
motor
cinemagoer
dnspython
imdbpie
python-dotenv
googletrans==3.1.0a0
bs4
Ponytile
PTL
